updated the post component to be more modular and broke it up into smaller pieces
